Fix test file name to build in CIwq
